backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
chromadb
pandas
pydantic
pydantic-settings
requests
